Fix for postgres

+++ games/models.py
@@ -7,5 +7,5 @@ class TalantUser(models.Model):
     access_token = models.CharField(max_length=2500)
 
     steam_openid = models.CharField(max_length=200, default='')
-    steam_id = models.IntegerField(default=0)
+    steam_id = models.BigIntegerField(default=0)
     # talant
